+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 * Support for using Encoder via context manager
 * Support for encoding sequences via context manager
 * Support for encoding sets via context manager
+* Support for using Decoder via context manager
 
 3.1.0 (2025-05-16)
 ------------------